Move ModuleTemplate tests and apply black
