+++ ceasiompy/ModuleTemplate/func/subfunc.py
@@ -39,7 +39,7 @@ log = get_logger(__file__.split(".")[0])
 
 
 def my_subfunc(arg_a, arg_b):
-    """ Function to clacluate ...
+    """Function to calculate ...
 
     Function 'my_subfunc' is a subfunction of ModuleTemplate which returns...
