--- test/TestModuleTemplate/ToolOutput/.keep
@@ -1,1 +1,0 @@
-
